Move function calling context to standalone module
